Makefile was in .gitignore, oops. Content plugin is much simpler now.
I am separating out the caching functionality from Content, and
ripped the old version out of Content in this release.
Also put in a quick fix for Headers with the new DateTime spec.
Oh, and the most embarrasing part, I had previously added Makefile
to the .gitignore file, and hadn't noticed that it was still there.
I removed it and added the custom Makefile that I introduced a few revisions
ago... oops!